refactor(agent): move stream types to agent-contract
