+++ src/agent-stream.ts
@@ -11,7 +11,7 @@ import type { Agent, StreamOptions, StreamOutput } from "./agent-contract";
 import { ERROR_KINDS, LIFECYCLE_ERROR_CODES } from "./error-contract";
 import { serializeToolError } from "./error-handling";
 import { MAX_TOOL_RESULT_CHARS } from "./lifecycle-constants";
-import type { GenerateResult, LifecycleSignal, StreamChunk, ToolCallEntry } from "./lifecycle-contract";
+import type { GenerateResult, LifecycleSignal, StreamChunk, ToolCallEntry } from "./agent-contract";
 import { log } from "./log";
 import { createModel } from "./model-factory";
 import { applyPromptCacheMarkers } from "./prompt-cache";
